* commit '7a03145ed7cb4f1ce794b5126559dd6f38029243':
  x86: dsputil: int --> ptrdiff_t for ff_put_pixels16_mmxext line_size param

Merged-by: Michael Niedermayer <michaelni@gmx.at>

+++ libavcodec/x86/dsputil_mmx.c
@@ -149,7 +149,7 @@ void ff_avg_pixels8_xy2_3dnow(uint8_t *block, const uint8_t *pixels,
 
 void ff_put_pixels8_mmxext(uint8_t *block, const uint8_t *pixels, ptrdiff_t line_size, int h);
 static void ff_put_pixels16_mmxext(uint8_t *block, const uint8_t *pixels,
-                                   int line_size, int h)
+                                   ptrdiff_t line_size, int h)
 {
     ff_put_pixels8_mmxext(block,     pixels,     line_size, h);
     ff_put_pixels8_mmxext(block + 8, pixels + 8, line_size, h);
